% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adapt_units}
\alias{adapt_units}
\title{Adapt units to a reference}
\usage{
adapt_units(x, ref)
}
\arguments{
\item{x}{input object.}

\item{ref}{an object of class \code{units}.}
}
\value{
An object of class \code{units}, with the same units as \code{ref}.
}
\description{
Helper function to adapt units to a given reference.
}
\examples{
x <- set_units(1000, m)
y <- set_units(5, km)
adapt_units(x, y)

}
