% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.hms.units}
\alias{as.hms.units}
\title{Convert units to hms}
\usage{
\method{as.hms}{units}(x, ...)
}
\arguments{
\item{x}{object of class units}

\item{...}{passed on to as.hms.difftime}
}
\value{
object of class hms
}
\description{
Convert units to hms
}
\examples{
if (require(hms)) {
 as.hms(1:10 * with(ud_units, s))
 as.hms(1:10 * with(ud_units, min))
 as.hms(1:10 * with(ud_units, h))
 as.hms(1:10 * with(ud_units, d))
}
}

