% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.units.difftime}
\alias{as.units.difftime}
\title{convert difftime objects to units}
\usage{
\method{as.units}{difftime}(x, value)
}
\arguments{
\item{x}{object of class units}

\item{value}{target unit; if omitted, taken from \code{x}}
}
\description{
convert difftime objects to units
}
\examples{
s = Sys.time()
d  = s - (s+1)
as.units(d)
}

