% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_cf}
\alias{as_cf}
\title{deparse unit to string in product power form (e.g. km m-2 s-1)}
\usage{
as_cf(x)
}
\arguments{
\item{x}{object of class units}
}
\description{
deparse unit to string in product power form (e.g. km m-2 s-1)
}
\examples{
u = parse_unit("kg m-2 s-1")
u
as_cf(u)
}
