% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_difftime}
\alias{as_difftime}
\title{convert units object into difftime object}
\usage{
as_difftime(x)
}
\arguments{
\item{x}{object of class \code{units}}
}
\description{
convert units object into difftime object
}
\examples{

t1 = Sys.time() 
t2 = t1 + 3600 
d = t2 - t1
du <- as_units(d)
dt = as_difftime(du)
class(dt)
dt
}
