% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot.units}
\alias{boxplot.units}
\title{Boxplot for \code{units} objects}
\usage{
\method{boxplot}{units}(x, ..., horizontal = FALSE)
}
\arguments{
\item{x}{object of class \code{units}, for which we want to plot the boxplot}

\item{...}{parameters passed on to \code{\link{boxplot.default}}}

\item{horizontal}{logical; whether boxplots should be horizontal;
default \code{FALSE} means vertical boxes.}
}
\description{
Boxplot for \code{units} objects
}
\examples{
units_options(parse = FALSE) # otherwise we break on the funny symbol!
u = set_units(rnorm(100), degree_C)
boxplot(u)
}
