% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert between units}
\usage{
convert_units(x, from, to)
}
\arguments{
\item{x}{numeric vector}

\item{from}{character; unit of \code{x}}

\item{to}{character; unit to convert \code{x} to}
}
\value{
numeric vector with \code{x} converted to new unit
}
\description{
Convert numeric vector from one unit to another compatible unit
}
\examples{
convert_units(100, "m", "km")
convert_units(32, "degF", "degC")
}
