% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{make_unit}
\alias{make_unit}
\title{deprecated functions}
\usage{
make_unit(x, ...)
}
\arguments{
\item{x}{see \link{symbolic_unit}}

\item{...}{see \link{symbolic_unit}}
}
\description{
The following functions are deprecated and will be removed in a future release.
}
