% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.units}
\alias{hist.units}
\title{Histogram for \code{units} objects}
\usage{
\method{hist}{units}(x, xlab = NULL, main = paste("Histogram of", xname),
  ...)
}
\arguments{
\item{x}{object of class \code{units}, for which we want to plot the histogram}

\item{xlab}{character; x axis label}

\item{main}{character; title of histogram}

\item{...}{parameters passed on to \code{\link{hist.default}}}
}
\description{
Histogram for \code{units} objects
}
\examples{
units_options(parse = FALSE) # otherwise we break on the funny symbol!
u = set_units(rnorm(100), degree_C)
hist(u)
}
