% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{matmult}
\alias{matmult}
\alias{\%*\%}
\alias{matmult}
\alias{\%*\%.default}
\alias{matmult}
\alias{\%*\%.units}
\title{matrix multiplication}
\usage{
x \%*\% y

\method{\%*\%}{default}(x, y)

\method{\%*\%}{units}(x, y)
}
\arguments{
\item{x}{numeric matrix or vector}

\item{y}{numeric matrix or vector}
}
\description{
matrix multiplication
}
\examples{
a = set_units(1:5, m)
a \%*\% a
a \%*\% t(a)
a \%*\% set_units(1:5, 1)
set_units(1:5, 1) \%*\% a
}
