% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{matrixOps.units}
\alias{matrixOps.units}
\title{S3 matrixOps Group Generic Functions for units objects}
\usage{
\method{matrixOps}{units}(x, y)
}
\arguments{
\item{x}{object of class \code{units},
or something that can be coerced to it by \code{as_units(x)}}

\item{y}{object of class \code{units},
or something that can be coerced to it by \code{as_units(y)}}
}
\value{
object of class \code{units}
}
\description{
matrixOps functions for units objects.
}
\examples{
a = set_units(1:5, m)
a \%*\% a
a \%*\% t(a)
a \%*\% 1:5
1:5 \%*\% a
}
