% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.units}
\alias{plot.units}
\title{plot unit objects}
\usage{
\method{plot}{units}(x, y, xlab = NULL, ylab = NULL, ..., sep = c(" [",
  "]"))
}
\arguments{
\item{x}{object of class units, to plot along the x axis, or, if y is missing, along the y axis}

\item{y}{object to plot along the y axis, or missing}

\item{xlab}{character; x axis label}

\item{ylab}{character; y axis label}

\item{...}{other parameters, passed on to \link{plot.default}}

\item{sep}{length two character vector; symbols to put before and after the measurement unit}
}
\description{
plot unit objects
}
\examples{
u = rnorm(1:10) * make_unit("°C")
v = rnorm(1:10) * make_unit("s")
plot(u,v)
plot(u, type = 'l')
}

