% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{plus}
\alias{plus}
\title{Title add two vectors with identical units}
\usage{
plus(x, y)
}
\arguments{
\item{x}{first vector, of class \code{units}}

\item{y}{second vector, of class \code{units}}
}
\value{
numeric vector of class \code{units}
}
\description{
Title add two vectors with identical units
}
\examples{
a = 1:3
b = 7:9
plus(a,b)
}

