% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{quantile}
\alias{quantile}
\title{quantile method for object of class units}
\usage{
quantile(x, ...)
}
\arguments{
\item{x}{object of class \code{units}}

\item{...}{arguments passed on to quantile}
}
\description{
quantile method for object of class units
}

