% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{seq.units}
\alias{seq.units}
\title{seq method for units objects}
\usage{
\method{seq}{units}(from, to, by = ((to - from)/(length.out - 1)),
  length.out = NULL, along.with = NULL, ...)
}
\arguments{
\item{from}{see \link[base]{seq}}

\item{to}{see \link[base]{seq}}

\item{by}{see \link[base]{seq}}

\item{length.out}{see \link[base]{seq}}

\item{along.with}{see \link[base]{seq}}

\item{...}{see \link[base]{seq}}
}
\description{
seq method for units objects
}
\details{
arguments with units are converted to have units of the first argument (which is either \code{from} or \code{to})
}
\examples{
seq(to = set_units(10, m), by = set_units(1, m), length.out = 5)
seq(set_units(10, m), by = set_units(1, m), length.out = 5)
seq(set_units(10, m), set_units(19, m))
seq(set_units(10, m), set_units(.1, km), set_units(10000, mm))
}
