% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_units.R
\name{set_units}
\alias{set_units}
\title{set_units}
\usage{
set_units(x, value, ..., mode = getOption("units.set_units_mode", c("symbols",
  "standard")))
}
\arguments{
\item{x}{a numeric to be assigned units, or a units object to have units
converted}

\item{value}{a \code{units} object, or something coercible to one with
\code{as_units}. Depending on \code{mode}, the unit is constructed from the
supplied bare expression or from the supplied value via standard
evaluation.}

\item{...}{passed on to \code{as_units}}

\item{mode}{if \code{"symbols"} (the default), then unit is constructed from
the expression supplied. Otherwise, if\code{mode = "standard"},
standard evaluation is used for the supplied value This argument can be set
via a global option \code{options(units.set_units_mode = "standard")}}
}
\description{
A pipe friendly version of \code{units<-}
}
\seealso{
\code{\link{as_units}}
}
