% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tibble}
\alias{tibble}
\alias{type_sum.units}
\alias{tibble}
\alias{pillar_shaft.units}
\title{type_sum function for units}
\usage{
type_sum.units(x, ...)

pillar_shaft.units(x, ...)
}
\arguments{
\item{x}{see \link[pillar]{type_sum}}

\item{...}{see \link[pillar]{type_sum}}
}
\description{
type_sum function for units

pillar_shaft function for units
}
