% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{type_sum.units}
\alias{type_sum.units}
\title{type_sum for tidy tibble printing}
\usage{
type_sum.units(x, ...)
}
\arguments{
\item{x}{object of class units}

\item{...}{ignored}
}
\description{
type_sum for tidy tibble printing
}

