% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udunits.R
\name{ud_convert}
\alias{ud_convert}
\title{Convert between units}
\usage{
ud_convert(x, from, to)
}
\arguments{
\item{x}{numeric vector}

\item{from}{character; unit of \code{x}}

\item{to}{character; unit to convert \code{x} to}
}
\value{
numeric vector with \code{x} converted to new unit
}
\description{
Convert numeric vector from one unit to another compatible unit
}
\examples{
ud_convert(100, "m", "km")
ud_convert(32, "degF", "degC")
}
