% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\docType{data}
\name{ud_units}
\alias{ud_units}
\title{Pre-defined units from UDUNITS2}
\format{
A list of class \code{units database}.
}
\usage{
ud_units
}
\description{
Lazy-loaded when used.
}
\keyword{datasets}
