% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{unit}
\alias{unit}
\title{convenience function to set units}
\usage{
unit(x, unit)
}
\arguments{
\item{x}{numeric vector}

\item{unit}{expression, representing units, e.g. m/s; will be evaluated in \link{ud_units}}
}
\description{
convenience function to set units
}
\examples{
unit(1:10, m/s)
unit(1:10, m^3)
}
