% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_units.R
\docType{data}
\name{unitless}
\alias{unitless}
\title{The "unit" type for vectors that are actually dimension-less.}
\format{An object of class \code{symbolic_units} of length 2.}
\usage{
unitless
}
\description{
The "unit" type for vectors that are actually dimension-less.
}
\keyword{datasets}
