% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{units-defunct}
\alias{units-defunct}
\title{Defunct functions in \pkg{units}}
\description{
These functions are no longer available.
}
\details{
\itemize{
 \item \code{ud_units}: Use \code{\link{as_units}} instead.
 \item \code{as.units}: Use \code{\link{as_units}} instead.
 \item \code{make_unit}: Use \code{\link{as_units}} instead.
 \item \code{parse_unit}: Use \code{\link{as_units}} instead.
 \item \code{as_cf}: Use \code{\link{deparse_unit}} instead.
 \item \code{install_symbolic_unit}: Use \code{\link{install_unit}} instead.
 \item \code{remove_symbolic_unit}: Use \code{\link{remove_unit}} instead.
 \item \code{install_conversion_constant}: Use \code{\link{install_unit}} instead.
 \item \code{install_conversion_offset}: Use \code{\link{install_unit}} instead.
}
}
