% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{units<-.numeric}
\alias{units<-.numeric}
\title{Set measurement units on a numeric vector}
\usage{
\method{units}{numeric}(x) <- value
}
\arguments{
\item{x}{numeric vector}

\item{value}{character; unit of measurement string}
}
\value{
object of class \code{units}
}
\description{
Set measurement units on a numeric vector
}
\examples{
x = 1:3
class(x)
units(x) <- with(ud_units, m/s) # valid
class(x)
y = 2:5
}

