% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{units<-}
\alias{units<-}
\title{Set measurement units on a numeric vector}
\usage{
units(x, strict.ud = FALSE) <- value
}
\arguments{
\item{x}{numeric vector}

\item{value}{character; unit of measurement string}

\item{strict.ud}{boolean; must the unit be valid as a udunit2 measurment?}
}
\value{
object of class \code{units}
}
\description{
Set measurement units on a numeric vector
}
\examples{
x = 1:3
class(x)
units(x) = "m/s" # valid
class(x)
y = 2:5
try(units(y) <- "xxy") # error
}

