% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{units}
\alias{units}
\alias{units<-.units}
\title{Convert units}
\usage{
\method{units}{units}(x) <- value
}
\arguments{
\item{x}{object of class \code{units}}

\item{value}{length one character vector with target unit}
}
\description{
Convert units
}
\examples{
a <- with(ud_units, 1:3 * m/s)
units(a) <- with(ud_units, km/h)
a
}

