% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{units.units}
\alias{units.units}
\title{retrieve measurement units from units object}
\usage{
\method{units}{units}(x)
}
\arguments{
\item{x}{object of class \code{units}}
}
\description{
retrieve measurement units from units object
}

