% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_conversion.R
\name{define_new_symbolic_unit}
\alias{define_new_symbolic_unit}
\alias{get_user_defined_units}
\alias{is_user_defined_unit}
\alias{remove_user_defined_symbolic_unit}
\title{Define new symbolic units}
\usage{
define_new_symbolic_unit(chr)

get_user_defined_units()

is_user_defined_unit(chr)

remove_user_defined_symbolic_unit(chr)
}
\description{
Adding a symbolic unit allows it to be used in \code{as_units}, \code{make_units} and \code{set_units}
}
