% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{as.dt}
\alias{as.dt}
\title{convert units object into difftime object}
\usage{
as.dt(x)
}
\arguments{
\item{x}{object of class \code{units}}
}
\description{
convert units object into difftime object
}
\examples{

t1 = Sys.time() 
t2 = t1 + 3600 
d = t2 - t1
as.units(d)
(du = as.units(d, "d"))
dt = as.dt(du)
class(dt)
dt
}

