% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as.units}
\alias{as.units}
\title{convert object to a units object}
\usage{
as.units(x, value = unitless)
}
\arguments{
\item{x}{object of class units}

\item{value}{target unit, defaults to `unitless`}
}
\description{
convert object to a units object
}

