% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot.units}
\alias{boxplot.units}
\title{boxplot for unit objects}
\usage{
\method{boxplot}{units}(x, ..., horizontal = FALSE)
}
\arguments{
\item{x}{object of class units, for which we want to plot the boxplot}

\item{...}{parameters passed on to \link{boxplot.default}}

\item{horizontal}{logical indicating if the boxplots should be horizontal; 
default FALSE means vertical boxes.}
}
\description{
boxplot for unit objects
}
\examples{
units_options(parse = FALSE) # otherwise we break on the funny symbol!
u = set_units(rnorm(100), degree_C)
boxplot(u)
}
