% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{deparse_unit}
\alias{deparse_unit}
\alias{deparse_unit}
\alias{as_cf}
\title{deparse unit to string in product power form (e.g. km m-2 s-1)}
\usage{
deparse_unit(x)

as_cf(x)
}
\arguments{
\item{x}{object of class units}
}
\value{
length one character vector
}
\description{
deparse unit to string in product power form (e.g. km m-2 s-1)
}
\details{
\code{as_cf} is deprecated; use \code{deparse_unit}.
}
\examples{
u = as_units("kg m-2 s-1", implicit_exponents = TRUE)
u
deparse_unit(u)
}
