% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{make_unit}
\alias{make_unit}
\alias{parse_unit}
\title{Deprecated functions}
\usage{
make_unit(chr, ...)

parse_unit(chr, ...)
}
\description{
The following functions are deprecated and will be removed in a future release.
}
