% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depracated.R
\name{make_unit}
\alias{make_unit}
\alias{parse_unit}
\alias{deprecated}
\alias{as.units}
\title{Deprecated functions}
\usage{
make_unit(chr)

parse_unit(chr)

as.units(x, value = unitless)
}
\arguments{
\item{chr}{length 1 character string}

\item{x}{a numeric}

\item{value}{a units object, by default, unitless}
}
\description{
The following function will be deprecated and removed in a future release.
}
