% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_units.R
\name{drop_units}
\alias{drop_units}
\title{drop units}
\usage{
drop_units(x)
}
\arguments{
\item{x}{a units object}
}
\value{
the numeric without any units attributes, while preserving other
  attributes like dimensions or other classes.
}
\description{
drop units
}
\note{
Equivalent to \code{units(x) <- NULL}
}
