% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.units}
\alias{hist.units}
\title{histogram for unit objects}
\usage{
\method{hist}{units}(x, xlab = NULL, ..., sep = c(" [", "]"))
}
\arguments{
\item{x}{object of class units, for which we want to plot the histogram}

\item{xlab}{x axis label}

\item{...}{parameters passed on to \link{hist.default}}

\item{sep}{length two character vector; symbols to put before and after the measurement unit}
}
\description{
histogram for unit objects
}
\examples{
u = rnorm(100) * make_unit("°C")
hist(u)
}

