% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.units}
\alias{hist.units}
\title{histogram for unit objects}
\usage{
\method{hist}{units}(x, xlab = NULL, main = paste("Histogram of", xname),
  ...)
}
\arguments{
\item{x}{object of class units, for which we want to plot the histogram}

\item{xlab}{character; x axis label}

\item{main}{character; title of histogram}

\item{...}{parameters passed on to \link{hist.default}}
}
\description{
histogram for unit objects
}
\examples{
units_options(parse = FALSE) # otherwise we break on the funny symbol!
u = rnorm(100) * make_unit("°C")
hist(u)
}
