% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_conversion.R
\name{install_symbolic_unit}
\alias{install_symbolic_unit}
\alias{uninstall_symbolic_unit}
\title{Define new symbolic units}
\usage{
install_symbolic_unit(chr, warn = TRUE)

uninstall_symbolic_unit(chr)
}
\arguments{
\item{chr}{a length 1 character vector that is the new unit name or symbol.}

\item{warn}{warns if the supplied unit symbol is already a valid unit symbol
recognized by udunits.}
}
\description{
Adding a symbolic unit allows it to be used in \code{as_units},
\code{make_units} and \code{set_units}. No installation is performed if the
unit is already known by udunits.
}
