% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_units.R
\name{make_unit}
\alias{make_unit}
\title{Create a new unit from a unit name.}
\usage{
make_unit(name)
}
\arguments{
\item{name}{Name of the new unit}
}
\value{
A new unit object that can be used in arithmetics
}
\description{
Create a new unit from a unit name.
}
