% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{Math.units}
\alias{Math.units}
\title{Mathematical operations for units objects}
\usage{
\method{Math}{units}(x, ...)
}
\arguments{
\item{x}{object of class units}

\item{...}{parameters passed on to the Math functions}
}
\description{
Mathematical operations for units objects
}
\examples{
a = sqrt(1:3)
units(a) = "m/s"
log(a)
cumsum(a)
signif(a, 2)
}

