% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_unit}
\alias{parse_unit}
\title{parse unit in product power form (e.g. km m-2 s-1)}
\usage{
parse_unit(str)
}
\arguments{
\item{str}{lenght-one character vector containing the unit string}
}
\description{
parse unit in product power form (e.g. km m-2 s-1)
}
\details{
see also \code{demo(cf)} for parsing units in the CF standard name table. Note that \code{parse_unit} currently fails on expressions containing a \code{/}, such as \code{m/s-1}.
}
\examples{
parse_unit("kg m-2 s-1")
}

