% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_units.R
\name{set_units}
\alias{set_units}
\alias{set_units.default}
\alias{set_units.difftime}
\title{set_units}
\usage{
set_units(x, ...)

\method{set_units}{default}(x, value, ...,
  mode = getOption("units.set_units_mode", c("symbols", "standard")))

\method{set_units}{difftime}(x, value, ...)
}
\arguments{
\item{x}{a numeric to be assigned units, or a units object to have units
converted}

\item{...}{passed on to \code{as_units}}

\item{value}{a \code{units} object, or something coercible to one with
\code{as_units}. Depending on \code{mode}, the unit is constructed from the
supplied bare expression or from the supplied value via standard evaluation.}

\item{mode}{if \code{"symbols"} (the default), then unit is constructed from
the bare expression supplied. Otherwise, if\code{mode = "standard"},
standard evaluation is used for the supplied value This argument can be set
via a global option \code{options(units.set_units_mode = "standard")}}
}
\description{
A pipe friendly version of \code{units<-}
}
\note{
for objects of class difftime, \code{set_units} is a simple alias for
  \code{`units<-`}, and calls \code{units(x) <- value} directly without
  calling \code{as_units(value)}
}
