% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/mixed.R
\name{tibble}
\alias{tibble}
\alias{type_sum.units}
\alias{tibble}
\alias{format_type_sum.type_sum_units}
\alias{tibble}
\alias{pillar_shaft.units}
\alias{tibble}
\alias{type_sum.mixed_units}
\alias{tibble}
\alias{pillar_shaft.mixed_units}
\title{type_sum function for units}
\usage{
type_sum.units(x, ...)

format_type_sum.type_sum_units(x, width, ...)

pillar_shaft.units(x, ...)

type_sum.mixed_units(x, ...)

pillar_shaft.mixed_units(x, ...)
}
\arguments{
\item{x}{see \link[pillar]{type_sum}}

\item{...}{see \link[pillar]{type_sum}}

\item{width}{ignored}
}
\description{
type_sum function for units

pillar_shaft function for units
}
