% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udunits.R
\name{ud_are_convertible}
\alias{ud_are_convertible}
\title{Test if two units are convertible}
\usage{
ud_are_convertible(u1, u2)
}
\arguments{
\item{u1}{String or object of class \code{symbolic_units}, for the symbol of the first unit.}

\item{u2}{String or object of class \code{symbolic_units}, for the symbol of the second unit.}
}
\value{
boolean, \code{TRUE} if both units are convertible.
}
\description{
Test if two units are convertible, though direct interface with UDUNITS-2.
}
\examples{
ud_are_convertible("m", "km")
a <- set_units(1:3, m/s)
ud_are_convertible(units(a), "km/h")
ud_are_convertible("s", "kg")
}
