% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udunits.R
\name{ud_are_convertible}
\alias{ud_are_convertible}
\title{Test if two units are convertible}
\usage{
ud_are_convertible(u1, u2)
}
\arguments{
\item{u1}{string, the first unit}

\item{u2}{string, the second unit}
}
\value{
boolean, \code{TRUE} if both units are convertible.
}
\description{
Test if two units are convertible
}
