% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ud_units.R
\docType{data}
\name{ud_units}
\alias{ud_units}
\title{List containing pre-defined units from the udunits2 package.}
\format{An object of class \code{NULL} of length 0.}
\usage{
ud_units
}
\description{
Lazy loaded when used
}
\keyword{datasets}
