% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R
\name{units<-.numeric}
\alias{units<-.numeric}
\title{Set measurement units on a numeric vector}
\usage{
\method{units}{numeric}(x) <- value
}
\arguments{
\item{x}{numeric vector}

\item{value}{character; valid unit of measurement string}
}
\value{
object of class \code{units}
}
\description{
Set measurement units on a numeric vector
}
\examples{
x = 1:3
class(x)
units(x) = "m/s" # valid
class(x)
y = 2:5
try(units(y) <- "xxy") # error
}

