% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{units}
\alias{units}
\alias{units<-.numeric}
\alias{units}
\alias{units<-.units}
\alias{units}
\alias{set_units}
\alias{units}
\alias{units.units}
\title{Set measurement units on a numeric vector}
\usage{
\method{units}{numeric}(x) <- value

\method{units}{units}(x) <- value

set_units(x, value)

\method{units}{units}(x)
}
\arguments{
\item{x}{numeric vector, or object of class \code{units}}

\item{value}{object of class \code{units} or \code{symbolic_units}}
}
\value{
object of class \code{units}

the units method retrieves the units attribute, which is of class \code{symbolic_units}
}
\description{
Set measurement units on a numeric vector

Convert units

retrieve measurement units from \code{units} object
}
\details{
\code{set_units} is a pipe-friendly version of \code{units<-}.
}
\examples{
x = 1:3
class(x)
units(x) <- with(ud_units, m/s) # valid
class(x)
y = 2:5
a <- with(ud_units, 1:3 * m/s)
units(a) <- with(ud_units, km/h)
a
if (require(magrittr)) {
 1:10 \%>\% set_units(with(ud_units, m)) \%>\% set_units(with(ud_units, km))
}
}
