% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_udunits.R
\name{valid_udunits}
\alias{valid_udunits}
\alias{valid_udunits}
\alias{valid_udunits_prefixes}
\title{Get information about valid units}
\usage{
valid_udunits(quiet = FALSE)

valid_udunits_prefixes(quiet = FALSE)
}
\value{
a data frame with columns `symbol`, `symbol_aliases`, `
  name_singular`, `name_singular_aliases`, `name_plural`, or
  `name_plural_aliases`, `def`, `definition`, `comment`, `dimensionless` and
  `source_xml`
}
\description{
The returned dataframe is constructed at runtime by reading the xml database
that powers unit conversion in [package:udunits2]. Inspect this dataframe to
determine what inputs are accepted by `as_units` (and the other functions it
powers: `make_units`, `set_units`, `units<-`). Any entry listed under
`symbol`, `symbol_aliases`, ` name_singular`, `name_singular_aliases`,
`name_plural`, or `name_plural_aliases` is valid. Additionaly, any entry
under `symbol` or `symbol_aliases` may can also contain a valid prefix, as,
as specified by `valid_udunits_prefixes()`.
}
\details{
Note, this is primarily intended for interactive use, the exact format of the
returned dataframe may change in the future.
}
\examples{
valid_udunits()
valid_udunits_prefixes()
if(interactive())
  View(valid_udunits())
}
